move pauli tracking functions to pauli_tracker module
